// hpmm-api/src/types/product.d.ts
export interface product {
  id_product: string;
  id_subcategory: string;
  nombre: string;
  descripcion: string;
  stock_actual: number;
  stock_maximo: number;
  fecha_vencimiento: Date;
  numero_lote: string;
  estado: boolean;
  created_at?: Date;
  updated_at?: Date;
}
export interface NewProduct extends Omit<product, "id_product"> {}